Apply Visual Studio full cleanup
This includes all the cleanup functions, except for:
- removing this.
- removing var
- removing/adding braces for one-line sentences

+++ MidiShapeShifter/MidiShapeShifter/Mss/SettingsFileInfo.cs
@@ -53,7 +53,7 @@ namespace MidiShapeShifter.Mss
         public override int GetHashCode()
         {
             int hash = 13;
-            hash = (hash * 7) + (int)this.Name.GetHashCode();
+            hash = (hash * 7) + this.Name.GetHashCode();
             hash = (hash * 7) + this.FileLocationType.GetHashCode();
             hash = (hash * 7) + this.FilePath.GetHashCode();
             return hash;
